refactor: usePlaylist
